Fix test failures.
I'm not sure where the dependency on 'six' comes froms (pytz, perhaps?).

And it looks like zc.buildout 1.5.2 changed its templates or something: witness
the BUILDOUT_ORIGINAL_PYTHONPATH changes, and the extra code in the generated
addsitepackages.  It's possible that my changes will break the tests if run
with zc.buildout 1.5.1.  If this happens, please yell at me, and I'll try to
come up with some ellipses to make them pass on both.